+++ src/libs/antares/study/include/antares/study/parts/thermal/cluster.h
@@ -34,6 +34,7 @@
 
 #include "../../fwd.h"
 #include "../common/cluster.h"
+#include "cost_provider.h"
 #include "defines.h"
 #include "ecoInput.h"
 #include "pollutant.h"
@@ -62,60 +63,6 @@ enum class LocalTSGenerationBehavior
     useGlobalParameter = 0,
     forceGen,
     forceNoGen
-};
-
-class CostProvider
-{
-public:
-    virtual ~CostProvider() = default;
-    virtual double getOperatingCost(uint serieIndex, uint hourInTheYear) const = 0;
-    virtual double getMarginalCost(uint serieIndex, uint hourInTheYear) const = 0;
-    virtual double getMarketBidCost(uint hourInTheYear, uint year) const = 0;
-};
-
-class ThermalCluster;
-
-class ConstantCostProvider: public CostProvider
-{
-public:
-    explicit ConstantCostProvider(const ThermalCluster& cluster);
-    virtual ~ConstantCostProvider() = default;
-    double getOperatingCost(uint serieIndex, uint hourInTheYear) const override;
-    double getMarginalCost(uint serieIndex, uint hourInTheYear) const override;
-    double getMarketBidCost(uint hourInTheYear, uint year) const override;
-
-private:
-    const ThermalCluster& cluster;
-};
-
-class ScenarizedCostProvider: public CostProvider
-{
-public:
-    explicit ScenarizedCostProvider(const ThermalCluster& cluster);
-    virtual ~ScenarizedCostProvider() = default;
-    double getOperatingCost(uint serieIndex, uint hourInTheYear) const override;
-    double getMarginalCost(uint serieIndex, uint hourInTheYear) const override;
-    double getMarketBidCost(uint hourInTheYear, uint year) const override;
-
-private:
-    /*!
-    ** \brief Production Cost, Market Bid Cost and Marginal Cost Matrixes - Per Hour and per Time
-    *Series
-    */
-    struct CostsTimeSeries
-    {
-        std::array<double, HOURS_PER_YEAR> productionCostTs;
-        std::array<double, HOURS_PER_YEAR> marketBidCostTS;
-        std::array<double, HOURS_PER_YEAR> marginalCostTS;
-    };
-
-    void resizeCostTS();
-    void ComputeMarketBidTS();
-    void MarginalCostEqualsMarketBid();
-    void ComputeProductionCostTS();
-
-    std::vector<CostsTimeSeries> costsTimeSeries;
-    const ThermalCluster& cluster;
 };
 
 double computeMarketBidCost(double fuelCost,
